Add type safety to themes array
