+++ src/models/variables.ts
@@ -1,31 +1,22 @@
-import { Dream } from "./types";
+// import { Dream } from "./types";
 
-export const themes = [
-  "teknikdrömmar",
-  "vardagsdrömmar",
-  "husdrömmar",
-  "sportdrömmar",
-  "resdrömmar",
-];
-
-export let name = "NAMN";
-export const dreams: Dream[] = [
-  {
-    id: 1,
-    name: "Lära mig HTML/CSS",
-    theme: "teknikdrömmar",
-    checked: true,
-  },
-  {
-    id: 2,
-    name: "Lära mig TypeScript",
-    theme: "teknikdrömmar",
-    checked: false,
-  },
-  {
-    id: 3,
-    name: "En dröm som tar flera rader lorem ipsum",
-    theme: "vardagsdrömmar",
-    checked: false,
-  },
-];
+// export const dreams: Dream[] = [
+//   {
+//     id: 1,
+//     name: "Lära mig HTML/CSS",
+//     theme: "teknikdrömmar",
+//     checked: true,
+//   },
+//   {
+//     id: 2,
+//     name: "Lära mig TypeScript",
+//     theme: "teknikdrömmar",
+//     checked: false,
+//   },
+//   {
+//     id: 3,
+//     name: "En dröm som tar flera rader lorem ipsum",
+//     theme: "vardagsdrömmar",
+//     checked: false,
+//   },
+// ];
